fix(monitor): center task board footer pill text
