+++ apps/harness-monitor/Tests/HarnessMonitorKitTests/HarnessMonitorActionTestSupport.swift
@@ -396,16 +396,6 @@ final class RecordingHarnessClient: HarnessMonitorClientProtocol, @unchecked Sen
   var timelineWindowCallCountsBySessionID: [String: Int] = [:]
   var acpTranscriptDelaysBySessionID: [String: Duration] = [:]
 
-  var calls: [Call] {
-    get { lock.withLock { callsStorage } }
-    set { lock.withLock { callsStorage = newValue } }
-  }
-
-  var detail: SessionDetail {
-    get { lock.withLock { detailStorage } }
-    set { lock.withLock { detailStorage = newValue } }
-  }
-
   init(detail: SessionDetail = PreviewFixtures.detail) {
     detailStorage = detail
   }
