+++ apps/harness-monitor/Tests/HarnessMonitorKitTests/TaskBoardHTTPRouteCatalogParityTests.swift
@@ -90,7 +90,7 @@ private func daemonHTTPPathConstants() throws -> [String: String] {
 
 private func swiftTaskBoardHTTPRoutes() throws -> [TaskBoardHTTPRoute] {
   let relativePaths = [
-    "apps/harness-monitor/Sources/HarnessMonitorKit/API/HarnessMonitorAPIClient+TaskBoard.swift",
+    "apps/harness-monitor/Sources/HarnessMonitorKit/API/HarnessMonitorAPIClient+TaskBoard.swift"
   ]
   let routes = try relativePaths.flatMap { relativePath in
     let source = try repoFileContents(relativePath: relativePath)
